+++ src/Services/Webauthn/AbstractValidatorFactory.php
@@ -4,6 +4,7 @@ namespace LaravelWebauthn\Services\Webauthn;
 
 use CBOR\Decoder;
 use Cose\Algorithm\Manager;
+use Cose\Algorithm\Signature;
 use CBOR\Tag\TagObjectManager;
 use Http\Adapter\Guzzle6\Client;
 use Webauthn\PublicKeyCredentialLoader;
@@ -40,22 +41,35 @@ abstract class AbstractValidatorFactory extends AbstractFactory
      */
     protected function getAttestationStatementSupportManager(Decoder $decoder) : AttestationStatementSupportManager
     {
-        $coseAlgorithmManager = new Manager();
-
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\ECDSA\ES256());
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\ECDSA\ES512());
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\EdDSA\EdDSA());
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\RSA\RS1());
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\RSA\RS256());
-        $coseAlgorithmManager->add(new \Cose\Algorithm\Signature\RSA\RS512());
-
         $attestationStatementSupportManager = new AttestationStatementSupportManager();
 
+        // https://www.w3.org/TR/webauthn/#none-attestation
         $attestationStatementSupportManager->add(new NoneAttestationStatementSupport());
+
+        // https://www.w3.org/TR/webauthn/#fido-u2f-attestation
         $attestationStatementSupportManager->add(new FidoU2FAttestationStatementSupport($decoder));
-        $attestationStatementSupportManager->add(new AndroidSafetyNetAttestationStatementSupport(new Client(), 'GOOGLE_SAFETYNET_API_KEY'));
+
+        // https://www.w3.org/TR/webauthn/#android-safetynet-attestation
+        if ($this->config->get('webauthn.google_safetynet_api_key', '') !== '') {
+            $attestationStatementSupportManager->add(new AndroidSafetyNetAttestationStatementSupport(new Client(), $this->config->get('webauthn.google_safetynet_api_key')));
+        }
+
+        // https://www.w3.org/TR/webauthn/#android-key-attestation
         $attestationStatementSupportManager->add(new AndroidKeyAttestationStatementSupport($decoder));
+
+        // https://www.w3.org/TR/webauthn/#tpm-attestation
         $attestationStatementSupportManager->add(new TPMAttestationStatementSupport());
+
+        // https://www.w3.org/TR/webauthn/#packed-attestation
+        $coseAlgorithmManager = new Manager();
+
+        $coseAlgorithmManager->add(new Signature\ECDSA\ES256());
+        $coseAlgorithmManager->add(new Signature\ECDSA\ES512());
+        $coseAlgorithmManager->add(new Signature\EdDSA\EdDSA());
+        $coseAlgorithmManager->add(new Signature\RSA\RS1());
+        $coseAlgorithmManager->add(new Signature\RSA\RS256());
+        $coseAlgorithmManager->add(new Signature\RSA\RS512());
+
         $attestationStatementSupportManager->add(new PackedAttestationStatementSupport($decoder, $coseAlgorithmManager));
 
         return $attestationStatementSupportManager;